+++ core/postgres.sh
@@ -1,0 +1,31 @@
+#!/bin/bash
+
+# define helpers
+$source_dir='https://raw.github.com/divio/osx-bootstrap/master'
+source $source_dir/core/helpers.sh
+
+# install postgres
+postgres=`brew list | grep postgres`
+if [[ ! $postgres ]]; then
+    echo '##### Installing Formula Postgres...'
+    
+    PYTHON=/usr/local/bin/python brew install postgres
+
+    # setup postgres
+    ln -sfv /usr/local/opt/postgresql/*.plist ~/Library/LaunchAgents
+    launchctl load ~/Library/LaunchAgents/homebrew.mxcl.postgresql.plist
+    initdb /usr/local/var/postgres -E utf8
+
+    # brew installs
+    brew install postgis
+
+    # brew fixes
+    brew unlink libxml2
+    
+    # setup postgres
+    createuser postgres -s
+
+    # setup postgis
+    createdb template_postgis
+    createlang plpgsql template_postgis
+fi