+++ CHANGELOG.rst
@@ -2,11 +2,22 @@
 OSX Bootstrap
 =============
 
+1.1.0
+-----
+- restructure to core modules
+- added template folder for setup files
+- added global helpers function script
+- added brew formula foreman
+- added extra heroku
+- added extra ruby
+- fixed extra php
+
 1.0.1
 -----
-- cleanup of /extras folder
+- cleanup of extras folder
 - fixed an issue with sudo on etc/resolver
 - added venv alias
+- added denv alias
 - added valentia studio in favor of mysqlworkbench and pgadmin3
 - added gem compass
 - added formula git-flow